Embed pages in float32 by default and show per-page progress
The first live page embedding in bfloat16 on CPU was slow enough to be
interrupted: most consumer CPUs emulate bfloat16. float32 needs about
twice the memory but runs natively; bfloat16 stays available.

+++ pipeline/docs_spike.py
@@ -119,7 +119,7 @@ class QwenVLEmbedder:
     """Qwen3-VL-Embedding through transformers, the way PixelRAG uses it."""
 
     def __init__(self, model_id: str = DEFAULT_MODEL, *, device: str = "cpu",
-                 dtype: str = "bfloat16", revision: str | None = None):
+                 dtype: str = "float32", revision: str | None = None):
         try:
             import torch
             from transformers import AutoProcessor, Qwen3VLForConditionalGeneration
